complete core APIs

+++ usecase/src/main/java/com/example/usecase/ITransactionService.java
@@ -1,7 +1,7 @@
 package com.example.usecase;
 
 public interface ITransactionService {
-    public TransactionResult Transfer(String fromAccountId, String toAccountId, Integer amount);
-    public TransactionResult Deposit(String accountId, Integer amount);
-    public TransactionResult Withdraw(String accountId, Integer amount);
+    public TransactionResult Transfer(String fromAccountId, String toAccountId, Double amount);
+    public TransactionResult Deposit(String accountId, Double amount);
+    public TransactionResult Withdraw(String accountId, Double amount);
 }
